+++ noodles-vcf/src/header/record/value/map/format/definition/v4_4.rs
@@ -5,30 +5,38 @@ use crate::{
 
 pub(super) fn definition(key: &str) -> Option<(Number, Type, &'static str)> {
     match key {
-        key::READ_DEPTHS => Some((Number::R, Type::Integer, "Read depth for each allele")),
+        key::READ_DEPTHS => Some((
+            Number::ReferenceAlternateBases,
+            Type::Integer,
+            "Read depth for each allele",
+        )),
         key::FORWARD_STRAND_READ_DEPTHS => Some((
-            Number::R,
+            Number::ReferenceAlternateBases,
             Type::Integer,
             "Read depth for each allele on the forward strand",
         )),
         key::REVERSE_STRAND_READ_DEPTHS => Some((
-            Number::R,
+            Number::ReferenceAlternateBases,
             Type::Integer,
             "Read depth for each allele on the reverse strand",
         )),
         key::READ_DEPTH => Some((Number::Count(1), Type::Integer, "Read depth")),
-        key::EXPECTED_ALTERNATE_ALLELE_COUNTS => {
-            Some((Number::A, Type::Integer, "Expected alternate allele counts"))
-        }
+        key::EXPECTED_ALTERNATE_ALLELE_COUNTS => Some((
+            Number::AlternateBases,
+            Type::Integer,
+            "Expected alternate allele counts",
+        )),
         key::FILTER => Some((
             Number::Count(1),
             Type::String,
             r#"Filter indicating if this genotype was "called""#,
         )),
-        key::GENOTYPE_LIKELIHOODS => Some((Number::G, Type::Float, "Genotype likelihoods")),
-        key::GENOTYPE_POSTERIOR_PROBABILITIES => {
-            Some((Number::G, Type::Float, "Genotype posterior probabilities"))
-        }
+        key::GENOTYPE_LIKELIHOODS => Some((Number::Samples, Type::Float, "Genotype likelihoods")),
+        key::GENOTYPE_POSTERIOR_PROBABILITIES => Some((
+            Number::Samples,
+            Type::Float,
+            "Genotype posterior probabilities",
+        )),
         key::CONDITIONAL_GENOTYPE_QUALITY => Some((
             Number::Count(1),
             Type::Integer,
@@ -38,12 +46,12 @@ pub(super) fn definition(key: &str) -> Option<(Number, Type, &'static str)> {
         key::HAPLOTYPE_QUALITY => Some((Number::Count(2), Type::Integer, "Haplotype quality")),
         key::MAPPING_QUALITY => Some((Number::Count(1), Type::Integer, "RMS mapping quality")),
         key::ROUNDED_GENOTYPE_LIKELIHOODS => Some((
-            Number::G,
+            Number::Samples,
             Type::Integer,
             "Phred-scaled genotype likelihoods rounded to the closest integer",
         )),
         key::ROUNDED_GENOTYPE_POSTERIOR_PROBABILITIES => Some((
-            Number::G,
+            Number::Samples,
             Type::Integer,
             "Phred-scaled genotype posterior probabilities rounded to the closest integer",
         )),
@@ -73,11 +81,13 @@ pub(super) fn definition(key: &str) -> Option<(Number, Type, &'static str)> {
             Type::Float,
             "Copy number genotype quality",
         )),
-        key::GENOTYPE_COPY_NUMBER_LIKELIHOODS => {
-            Some((Number::G, Type::Float, "Copy number genotype likelihood"))
-        }
+        key::GENOTYPE_COPY_NUMBER_LIKELIHOODS => Some((
+            Number::Samples,
+            Type::Float,
+            "Copy number genotype likelihood",
+        )),
         key::GENOTYPE_COPY_NUMBER_POSTERIOR_PROBABILITIES => Some((
-            Number::G,
+            Number::Samples,
             Type::Float,
             "Copy number posterior probabilities",
         )),