+++ noodles-vcf/src/io/writer/header/record/value/map/number.rs
@@ -13,9 +13,9 @@ where
 
     match number {
         Number::Count(n) => write!(writer, "{n}"),
-        Number::A => writer.write_all(A),
-        Number::R => writer.write_all(R),
-        Number::G => writer.write_all(G),
+        Number::AlternateBases => writer.write_all(A),
+        Number::ReferenceAlternateBases => writer.write_all(R),
+        Number::Samples => writer.write_all(G),
         Number::Unknown => writer.write_all(UNKNOWN),
     }
 }
@@ -36,9 +36,9 @@ mod tests {
         let mut buf = Vec::new();
 
         t(&mut buf, Number::Count(1), b"1")?;
-        t(&mut buf, Number::A, b"A")?;
-        t(&mut buf, Number::R, b"R")?;
-        t(&mut buf, Number::G, b"G")?;
+        t(&mut buf, Number::AlternateBases, b"A")?;
+        t(&mut buf, Number::ReferenceAlternateBases, b"R")?;
+        t(&mut buf, Number::Samples, b"G")?;
         t(&mut buf, Number::Unknown, b".")?;
 
         Ok(())